fix: OpenCV-5 WHOLE_FACE affine float32 compatibility
cv2.getAffineTransform in OpenCV 5.x enforces CV_32F inputs,
while the official remove_align branch of get_transform_mat
(the WHOLE_FACE path) builds l_t from Python float lists
(float64); on OpenCV 5 the call raised cv2.error. The OpenCV
4.x python bindings downcast automatically, so the explicit
l_t.astype(np.float32) at the call site reproduces the official
behavior identically on both major versions (behavior-preserving
compatibility fix; no numerics change).

Pre-existing official-code x OpenCV-5 compatibility defect,
surfaced by the Phase 6B real-artifact acceptance (WHOLE_FACE
faceset sample processing); NOT a Phase 6B regression, kept as
its own logical fix.

Verified: a 60s CPU probe on real faceset data
(get_transform_mat WHOLE_FACE / FULL_NO_ALIGN -> finite 2x3
matrices, no cv2.error) and the real SAEHD one-step training
lifecycle on CUDA, which exercises this path end-to-end.

+++ facelib/LandmarksProcessor.py
@@ -346,7 +346,12 @@ def get_transform_mat (image_landmarks, output_size, face_type, scale=1.0):
 
     # calc affine transform from 3 global space points to 3 local space points size of 'output_size'
     pts2 = np.float32(( (0,0),(output_size,0),(output_size,output_size) ))
-    mat = cv2.getAffineTransform(l_t,pts2)
+    # OpenCV 5.x enforces CV_32F in getAffineTransform while the
+    # remove_align branch builds l_t from Python float lists
+    # (g_c + [..] -> float64); the explicit downcast reproduces what
+    # the OpenCV 4.x python bindings applied automatically, so the
+    # official behavior is identical on both major versions.
+    mat = cv2.getAffineTransform(l_t.astype(np.float32),pts2)
     return mat
 
 def get_rect_from_landmarks(image_landmarks):
